refactor(freeform-text): fix the css and refactor some functions as hook
